# .worktrees/main-hotfix/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Today’s Daily Battle — HTTP API (pilot data)",
    "version": "0.1.0",
    "description": "Read-only JSON for daily verse, battle plans, and calm-mood library. Primary site remains static/Next export; this service is for programmatic access (CORS enabled)."
  },
  "servers": [{ "url": "http://localhost:8787", "description": "Local" }],
...
  "paths": {
...
    "/v1/plans/{slug}": {
      "get": {
        "tags": ["Plans"],
        "summary": "Single battle plan by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}